requirements.txt
graphql-core>=2.1
dataclasses; python_version == '3.6'
typing-inspect
